moving sensor and binary sensor to packages
